+++ findingaids/fa/views.py
@@ -522,9 +522,13 @@ def search(request):
 
             # optional filter: restrict to items with digital archival objects
             if dao:
-                findingaids = findingaids.filter(daos__exists=True) \
-                                         .filter(public_dao_count__gte=1)
-                # restrict to public daos for now
+                findingaids = findingaids.filter(daos__exists=True)
+
+                # if user does not have permission to view internal daos,
+                # restrict to public daos only
+                if not request.user.has_perm('fa_admin.can_view_internal_dao'):
+                    findingaids = findingaids.filter(public_dao_count__gte=1)
+
                 # NOTE: using >= filter to force a where clause because this works
                 # when what seems to be the same filter on the xpath does not
                 # (possibly an indexing issue?)
@@ -644,9 +648,11 @@ def document_search(request, id):
 
             # restrict to publicly-accessible dao items, if set
             if form.cleaned_data['dao']:
-                files = files.filter(did__dao_list__exists=True) \
-                             .or_filter(did__dao_list__audience='external',
-                                        did__dao_list__audience__exists=False)
+                files = files.filter(did__dao_list__exists=True)
+                # if user can view internal daos, no additional filter is needed
+                # otherwise, restrict to publicly-accessible dao content
+                if not request.user.has_perm('fa_admin.can_view_internal_dao'):
+                    files = files.filter(public_dao_count__gte=1)
 
             files = files.also('parent__id', 'parent__did',
                                'series1__id', 'series1__did', 'series2__id', 'series2__did')
@@ -661,6 +667,7 @@ def document_search(request, id):
                 'ead': ead,
                 'querytime': [files.queryTime(), ead.queryTime()],
                 'keywords': search_terms,
+                'dao': form.cleaned_data['dao'],
                 'url_params': url_params,
                 'docsearch_form': KeywordSearchForm(),
             }, context_instance=RequestContext(request))